Move to directory structure for bokeh serve. Layout broken at the moment.
