+++ utilities/dashboard/app/general_barplot.py
