+++ utilities/dashboard/app/static/js/bootstrap.js
