+++ utilities/dashboard/app/static/js/fontawesome.js
