+++ utilities/dashboard/app/static/js/theme.js
