+++ utilities/dashboard/app/templates/index.html
@@ -31,10 +31,9 @@
         }
       </style>
 
-      <style type="text/css">{% include '_static/css/bootstrap.css' %}</style>
-      <style type="text/css">{% include '_static/css/theme.css' %}</style>
-      <style type="text/css">{% include '_static/css/custom.css' %}</style>
-
+      <link rel="stylesheet" ref="app/static/css/bootstrap.css">
+      <link rel="stylesheet" ref="app/static/css/theme.css">
+      <link rel="stylesheet" ref="app/static/css/custom.css">
       <link rel="stylesheet" type="text/css" href="https://assets.readthedocs.org/static/css/badge_only.css" />
 
       <meta name="title" content="CoViD-19 Impact Lab" />
@@ -52,7 +51,7 @@
 
       {{ bokeh_css }}
       {{ bokeh_js }}
-      <style type="text/css">{% include 'styles.css' %}</style>
+      <link rel="stylesheet" ref="app/static/css/styles.css">
     </head>
     <!-- =========================================================================== -->
     <!-- head ends-->
