+++ utilities/dashboard/app_mwe/styles.css
@@ -1,0 +1,9 @@
+button.bk.bk-btn.bk-btn-default {
+    color: red;
+    font-size:12pt;
+    background-color: blue;
+    border-color: green;
+}
+
+
+
